Moving observation classes + tests to new file, observations.py
