+++ gammapy/data/__init__.py
@@ -16,3 +16,4 @@ from .observers import *
 from .obs_table import *
 from .obs_summary import *
 from .obs_stats import *
+from .observations import *
